+++ contracts/near/eth-client/rust-toolchain
@@ -1,1 +1,1 @@
-stable-2020-10-08
+nightly-2020-10-08
